+++ audit/forgerock-audit-handler-jms/src/test/java/org/forgerock/audit/handlers/jms/JmsAuditEventHandlerTest.java
@@ -193,8 +193,8 @@ public class JmsAuditEventHandlerTest {
 
         // The first message gets picked up immediately, then the rest should be delivered in a 2 new batches as
         // the events remaining will be 1 more than the maxBatchedEvents.
-        // We should then expect only 3 sessions.
-        verify(connection, times(3)).createSession(anyBoolean(), anyInt());
+        // We should then expect 3 sessions, unless the threads are really fast and only 2 are needed.
+        verify(connection, atMost(3)).createSession(anyBoolean(), anyInt());
         // verify the total count of messages sent.
         verify(producer, times(messagesToSend)).send(any(TextMessage.class));
     }