+++ audit/CHANGELOG.md
@@ -6,6 +6,7 @@
 * CAUD-331: Disabled handlers should not utilize services until enabled.
 * CAUD-332: Tibco support for JMS Audit Handler: use JMS 2.0 dependency, and address JNDI lookup classpath issues.
 * CAUD-333: NPE when connection wasn't established with unhelpful error messages.
+* CAUD-335: Repair unstable JmsAuditEventHandlerTest and RotatableWriterTest.
 
 # 6.1.0